fix: service app panes that are not rendering (stint 0688)
Every assistant tool call to an app pane that was not painting timed out
after 30s. Reproduced on plain alpha with a fixture app predating the
report, so this is not a regression from any feature branch — it is how
app-pane servicing has always worked.

A tool call never touches the render path on either side: tool_dispatch
writes the ToolCall straight into the guest's stdin and blocks on a
channel, and the guest answers on stdout. The guest ran and replied
correctly every time. The host simply never read the reply.

It failed on two independent axes at once.

Occluded window. The whole app-command drain — drain_all_app_commands and
the ~1150-line execution loop it feeds — lived in App::ui. eframe skips ui
entirely for a minimized or fully covered window, so such a host executed
no app command of any runtime: no tool results, no notifications, no host
requests. That is the trap the root AGENTS.md already documents for pane
IPC, screenshots and the event bus; app panes are external clients by the
same definition and were the one drain still in ui. Extracted verbatim as
service_app_commands and called from App::logic. Only the modal-render
follow-up stays in ui, because early_modal_cmds is produced by that pass's
overlay draw.

Inactive context. AppRuntime::Python::background_tick was an empty arm and
needs_background_tick returned false, so the headless tick that keeps
Builtin and Wasm panes alive off screen skipped Python panes even on
visible passes. LivePythonPane now drains its decoder from background_tick.
It deliberately does not fire timers: those push onto pending_timer_events,
flushed only by the render request in ui, so firing them unpainted would
grow that queue without bound behind a repeating timer.

needs_background_tick is asked of every off-screen pane every frame and
must not consume, but mpsc::Receiver has no length — so the decoder counts
outputs it sends and PythonOutputDecoder::try_recv decrements. Counting
before the send, never after, keeps the paint thread from decrementing
past an increment and wrapping the unsigned counter to "always busy".

Guards, both verified falsifiable against the pre-fix code:
app_commands_execute_while_window_hidden (fails with the drain back in ui)
and python_tool_result_reaches_the_host_without_a_ui_pass, which launches
a real guest, renders one visible pass to let it expose its tools, then
services it through background_tick alone — 70s timeout before, 9.7s pass
after.
