+++ AGENTS.md
@@ -133,6 +133,7 @@ Non-obvious discoveries with no single owning directory. When you discover a tra
 - **egui advances delayed repaint requests by its predicted frame time.** A scheduler that passes its exact frame interval to `request_repaint_after` can turn a 60 Hz deadline into an immediate repaint loop. Add `InputState::predicted_dt` to the remaining wall-clock delay, and stop scheduling a deadline while the producer is still busy with an overdue frame. A literal zero-delay request also schedules an extra settling paint; use a nonzero delay when a background completion needs exactly one immediate host pass.
 - **macOS App Nap silently defers an idle host's cross-thread wakeups.** A non-frontmost idle eframe host gets napped: `request_repaint()` posted from a background thread (notify-socket IPC) is not delivered until an unrelated event un-naps the app, so CLI requests time out and then drain in one late burst. Neither winit nor eframe exempts the process. The host holds a process-lifetime `NSProcessInfo` activity (`platform::app_nap::disable_app_nap`, `UserInitiatedAllowingIdleSystemSleep`) so it stays wakeable; any new long-lived headless/background process serving IPC on macOS needs the same exemption.
 - **eframe skips `App::ui` entirely while the window is hidden.** `ViewportInfo::visible()` is false when the window is minimized *or fully occluded by other windows*; eframe then runs logic-only passes — `App::logic` fires, `App::ui` never does, and the pass still advances `cumulative_pass_nr`, so wakes "work" while nothing visible in `ui` executes. Every drain that services external clients (pane IPC, spawn queue, PTY events, shutdown, event-bus subscriptions, screenshot capture) must live in `App::logic`, never in `ui`. Regression guard: `pane_ipc_serviced_while_window_hidden`; harness pump: `HostHarness::hidden_frame`. This also applies to *issuing* work, not just draining it: `plexi host screenshot` sent its `ViewportCommand::Screenshot` from `ui`, so an occluded host never asked for a capture — the PNG arrived only once the window was uncovered, after the CLI had timed out (`service_pending_screenshots`, stints 0495/0504).
+- **"Not rendering" has two independent axes, and app panes are external clients on both.** An app answers assistant tool calls, raises notifications, and requests host actions on its own schedule — none of it caused by being painted — so app-pane servicing is external-client servicing and belongs in `App::logic` like every other drain above. It failed on *both* axes at once (stint 0688): the whole app-command drain lived in `App::ui`, so a minimized or occluded host executed no app command of any runtime; and `AppRuntime::Python`'s `background_tick` was an empty arm with `needs_background_tick` hardcoded `false`, so a Python pane in an inactive context was never serviced even on visible passes. A tool call reaches a guest through its stdin and is answered on its stdout with no render on either side, so both failures presented identically: the guest replied correctly and the host never read it, until `tool_dispatch`'s 30s timeout. When adding a runtime or a servicing path, satisfy both axes — `service_app_commands` for the occluded-window axis, `AppRuntime::background_tick`/`needs_background_tick` for the inactive-context axis. Guards: `app_commands_execute_while_window_hidden`, `python_tool_result_reaches_the_host_without_a_ui_pass`.
 - **`rustfmt` on a mod-declaring file reformats every reachable module.** `rustfmt src/main.rs` rewrites dozens of unrelated files because rustfmt follows `mod` declarations. Alpha is not fmt-clean, so this pollutes the diff exactly like bare `cargo fmt`. Format per file with `rustfmt --check` and hand-apply, or format only leaf files you actually touched.
 
 ## Architecture
